+++ src/main/java/ru/abs/bitrixdocbot/bitrix/BitrixRestClient.java
@@ -115,7 +115,7 @@ public class BitrixRestClient {
             byte[] data = restClientBuilder.build()
                 .get()
                 .uri(url)
-                .header(HttpHeaders.USER_AGENT, "bitrix-document-ai-bot/0.1.9")
+                .header(HttpHeaders.USER_AGENT, "bitrix-document-ai-bot/0.1.10")
                 .accept(MediaType.APPLICATION_OCTET_STREAM, MediaType.ALL)
                 .retrieve()
                 .onStatus(status -> status.isError(), (httpRequest, httpResponse) -> {
@@ -185,7 +185,7 @@ public class BitrixRestClient {
             RestClient.RequestBodySpec request = restClientBuilder.build()
                 .post()
                 .uri(url)
-                .header(HttpHeaders.USER_AGENT, "bitrix-document-ai-bot/0.1.9")
+                .header(HttpHeaders.USER_AGENT, "bitrix-document-ai-bot/0.1.10")
                 .accept(MediaType.APPLICATION_OCTET_STREAM, MediaType.ALL);
 
             byte[] data = requestCustomizer.apply(request)